[Core] Move ExecuteHelper to a more sensible location
