+++ src/parser/monk/shared/modules/spells/TouchOfDeath.js
@@ -2,7 +2,7 @@ import { SELECTED_PLAYER } from 'parser/core/Analyzer';
 import Abilities from 'parser/core/modules/Abilities';
 import SPELLS from 'common/SPELLS';
 import Events from 'parser/core/Events';
-import ExecuteHelper from 'parser/shared/ExecuteHelper';
+import ExecuteHelper from 'parser/shared/modules/helpers/ExecuteHelper';
 
 const MINUTE = 600000;
