+++ src/parser/paladin/shared/spells/HammerOfWrath.tsx
@@ -2,7 +2,7 @@ import { Options, SELECTED_PLAYER } from 'parser/core/Analyzer';
 import Abilities from 'parser/core/modules/Abilities';
 import SPELLS from 'common/SPELLS';
 import Events, { FightEndEvent } from 'parser/core/Events';
-import ExecuteHelper from 'parser/shared/ExecuteHelper';
+import ExecuteHelper from 'parser/shared/modules/helpers/ExecuteHelper';
 import Spell from 'common/SPELLS/Spell';
 
 class HammerofWrath extends ExecuteHelper {
